% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvcure.R
\name{expand.inits.args}
\alias{expand.inits.args}
\title{Taken from flexsurv, needed to wrap init functions of base distributions}
\usage{
expand.inits.args(inits)
}
\description{
Taken from flexsurv, needed to wrap init functions of base distributions
}
