% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.r
\name{mixsurv}
\alias{mixsurv}
\alias{dmixsurv}
\alias{pmixsurv}
\alias{qmixsurv}
\alias{rmixsurv}
\alias{hmixsurv}
\alias{Hmixsurv}
\alias{mean_mixsurv}
\alias{rmst_mixsurv}
\title{Mixture cure models}
\arguments{
\item{pfun}{The base distribution's cumulative distribution function.}

\item{dfun}{The base distribution's probability density function.}

\item{x, q, t}{Vector of times.}

\item{x, q, t}{Vector of times.}

\item{p}{Vector of probabilities.}

\item{n}{Number of random numbers to simulate.}

\item{theta}{The estimated cure fraction.}

\item{...}{Parameters to be passed to the pdf or cdf of the base
distribution.}

\item{start}{Optional left-truncation time or times.  The returned
restricted mean survival will be conditioned on survival up to
this time.}

\item{log, log.p}{Return log density or probability.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X
\le x)}{P(X <= x)}, otherwise, \eqn{P(X > x)}{P(X > x)}.}
}
\value{
\code{dmixsurv} gives the density, \code{pmixsurv} gives the
distribution function, \code{hmixsurv} gives the hazard and
\code{Hmixsurv} gives the cumulative hazard.

\code{qmixsurv} gives the quantile function, which is computed by crude
numerical inversion.

\code{rmixsurv} generates random survival times by using
\code{qmmixsurv} on a sample of uniform random numbers.  Due to the
numerical root-finding involved in \code{qmmixsurv}, it is slow compared
to typical random number generation functions.
}
\description{
Probability density, distribution, quantile, random generation, hazard
cumulative hazard, mean and restricted mean functions for generic
mixture cure models
}
\seealso{
\code{\link{flexsurvspline}}.
}
\author{
Jordan Amdahl <jrdnmdhl@gmail.com>
}
\keyword{distribution}
