% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmixture.R
\name{nmixsurv}
\alias{nmixsurv}
\alias{dnmixsurv}
\alias{pnmixsurv}
\alias{qnmixsurv}
\alias{rnmixsurv}
\alias{hnmixsurv}
\alias{Hnmixsurv}
\alias{mean_nmixsurv}
\alias{rmst_nmixsurv}
\title{Non-Mixture cure models}
\arguments{
\item{pfun}{The base distribution's cumulative distribution function.}

\item{dfun}{The base distribution's probability density function.}

\item{x, q, t}{Vector of times.}

\item{x, q, t}{Vector of times.}

\item{p}{Vector of probabilities.}

\item{n}{Number of random numbers to simulate.}

\item{theta}{The estimated cure fraction.}

\item{...}{Parameters to be passed to the pdf or cdf of the base
distribution.}

\item{start}{Optional left-truncation time or times.  The returned
restricted mean survival will be conditioned on survival up to
this time.}

\item{log, log.p}{Return log density or probability.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X
\le x)}{P(X <= x)}, otherwise, \eqn{P(X > x)}{P(X > x)}.}
}
\value{
\code{dnmixsurv} gives the density, \code{pnmixsurv} gives the
distribution function, \code{hnmixsurv} gives the hazard and
\code{Hnmixsurv} gives the cumulative hazard.

\code{qnmixsurv} gives the quantile function, which is computed by crude
numerical inversion.

\code{rnmixsurv} generates random survival times by using
\code{qnmixsurv} on a sample of uniform random numbers.  Due to the
numerical root-finding involved in \code{qnmixsurv}, it is slow compared
to typical random number generation functions.
}
\description{
Probability density, distribution, quantile, random generation, hazard
cumulative hazard, mean and restricted mean functions for generic
non-mixture cure models
}
\author{
Jordan Amdahl <jrdnmdhl@gmail.com>
}
\keyword{distribution}
